ebs+s3 without trival move

+++ cloud/cloud_storage_provider.cc
@@ -4,7 +4,7 @@
 #include "rocksdb/cloud/cloud_storage_provider.h"
 
 #include <cinttypes>
-
+#include <cstdio>
 #include "cloud/filename.h"
 #include "file/filename.h"
 #include "rocksdb/cloud/cloud_file_system.h"
@@ -155,6 +155,10 @@ CloudStorageWritableFileImpl::~CloudStorageWritableFileImpl() {
   }
 }
 
+static bool isEBS(std::string fname) {
+  return fname.find("ebs") != std::string::npos;
+}
+
 IOStatus CloudStorageWritableFileImpl::Close(const IOOptions& opts,
                                              IODebugContext* dbg) {
   if (local_file_ == nullptr) {  // already closed
@@ -180,11 +184,13 @@ IOStatus CloudStorageWritableFileImpl::Close(const IOOptions& opts,
       Log(InfoLogLevel::ERROR_LEVEL, cfs_->GetLogger(),
           "[%s] CloudWritableFile closing PutObject failed on local file %s",
           Name(), fname_.c_str());
+      printf("ERROR: CopyLocalFile %s ToDest %s failed\n", fname_.c_str(), cloud_fname_.c_str());
       return status_;
     }
 
     // delete local file
-    if (!cfs_->GetCloudFileSystemOptions().keep_local_sst_files) {
+    bool is_ebs = isEBS(fname_);
+    if (!cfs_->GetCloudFileSystemOptions().keep_local_sst_files && !is_ebs) {
       status_ = cfs_->GetBaseFileSystem()->DeleteFile(fname_, opts, dbg);
       if (!status_.ok()) {
         Log(InfoLogLevel::ERROR_LEVEL, cfs_->GetLogger(),
